requirements.txt
google-adk>=0.5.0
google-generativeai>=0.8.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
pydantic>=2.0.0
mcp>=1.0.0
